refactor: input -> field
