+++ packages/core/src/components/go-field/go-field.scss
@@ -1,0 +1,3 @@
+go-field {
+ display: block;
+}
