+++ packages/core/src/components/go-field/readme.md
@@ -1,0 +1,21 @@
+---
+title: Field
+---
+
+# Field <span class="text-size-0">`go-field`</span>
+
+<!-- Description -->
+
+<div class="text-size-1">
+  Field is a component in Go UI.
+</div>
+
+
+## Accessibility
+
+
+## Demo
+
+<demo-frame component="go-field" demo="go-field"></demo-frame>
+
+<!-- Auto Generated Below -->
