+++ packages/core/src/components/go-field/test/go-field.spec.tsx
@@ -1,0 +1,11 @@
+import { newSpecPage } from '@stencil/core/testing';
+import { GoField } from '../go-field';
+describe('go-field', () => {
+ it('renders', async () => {
+   const page = await newSpecPage({
+     components: [GoField],
+     html: `<go-field></go-field>`,
+   });
+   expect(page.root).toBeTruthy();
+ });
+});
