+++ packages/core/src/components/go-field/usage/go-field.md
@@ -1,0 +1,2 @@
+
+<go-field></go-field>
