+++ packages/core/src/index.html
@@ -35,33 +35,43 @@
 
       <!-- #region demo start -->
       <go-input id="test" type="text" label="Input field" hint="Please enter some text"></go-input>
-<!-- <go-input type="text" label="Error state" hint="Please enter some text" error="This is an error"></go-input>
-<go-input type="text" label="Disabled state" value="I'm not editable" disabled></go-input>
-<go-input type="text" label="Readonly state" value="I'm not editable" readonly></go-input>
-<go-input type="text" label="Custom icon">
-  <go-icon icon-set="material-icons" name="search" slot="icon-before"></go-icon>
-  <go-icon icon-set="material-icons" name="star_outline" slot="icon-after"></go-icon>
-</go-input>
-<go-input type="text" label="Prefix slot" hint="icon-before won't work because prefix is present">
-  <div slot="prefix">https://go-ui.com/</div>
+       <go-input type="text" label="Error state" hint="Please enter some text" error="This is an error"></go-input>
+      <go-input type="text" label="Disabled state" value="I'm not editable" disabled></go-input>
+      <go-input type="text" label="Disabled state" value="I'm not editable" disabled="disabled"></go-input>
+      <go-input type="text" label="Readonly state" value="I'm not editable" readonly></go-input>
+      <go-input type="text" label="Custom icon">
+        <go-icon icon-set="material-icons" name="search" slot="icon-before"></go-icon>
+        <go-icon icon-set="material-icons" name="star_outline" slot="icon-after"></go-icon>
+      </go-input>
 
-  <go-icon icon-set="material-icons" name="search" slot="icon-before"></go-icon>
-  <go-icon icon-set="material-icons" name="star_outline" slot="icon-after"></go-icon>
-</go-input>
-<go-input type="text" label="Suffix slot" hint="icon-after won't work because suffix is present">
+      <go-input type="text" label="Prefix slot">
+        <div slot="prefix">https://go-ui.com/</div>
+
+        <go-icon icon-set="material-icons" name="search" slot="icon-before"></go-icon>
+        <go-icon icon-set="material-icons" name="star_outline" slot="icon-after"></go-icon>
+      </go-input>
+      <go-input type="text" label="Suffix slot">
+        <div slot="suffix">.com.au</div>
+
+        <go-icon icon-set="material-icons" name="search" slot="icon-before"></go-icon>
+        <go-icon icon-set="material-icons" name="star_outline" slot="icon-after"></go-icon>
+      </go-input>
+
+<hr >
+
+<go-select label="Select">
+  <option value="1">Apple</option>
+  <option value="2">Banana</option>
+</go-select>
+
+<go-select label="Suffix slot" hint="icon-after won't work because suffix is present">
+  <option value="1">Apple</option>
+  <option value="2">Banana</option>
   <div slot="suffix">.com.au</div>
 
   <go-icon icon-set="material-icons" name="search" slot="icon-before"></go-icon>
   <go-icon icon-set="material-icons" name="star_outline" slot="icon-after"></go-icon>
-</go-input> -->
-
-<hr >
-
-<go-select label="Select" options="[
-{name: 'Hello', value: 'hello'},
-{name: 'World', value: 'world'}
-]"></go-select>
-
+</go-select>
       <!-- #endregion demo finish -->
     </div>
   </body>
